# remove a dependency, do not load SASS every time this gem is required