--- a/lib/zendesk_apps_support.rb
+++ b/lib/zendesk_apps_support.rb
@@ -1,5 +1,4 @@
 module ZendeskAppsSupport
-  require 'zendesk_apps_support/sass_functions'
   require 'zendesk_apps_support/engine'
 
   autoload :AppFile,                'zendesk_apps_support/app_file'