--- a/lib/zendesk_apps_support/manifest/no_override_hash.rb
+++ b/lib/zendesk_apps_support/manifest/no_override_hash.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
-require 'json'
-
 module ZendeskAppsSupport
   class Manifest
     class OverrideError < StandardError