--- a/lib/zendesk_apps_support/stylesheet_compiler.rb
+++ b/lib/zendesk_apps_support/stylesheet_compiler.rb
@@ -1,4 +1,5 @@
 require 'sass'
+require 'zendesk_apps_support/sass_functions'
 
 module ZendeskAppsSupport
   class StylesheetCompiler