--- a/lib/zendesk_apps_support/validations/manifest.rb
+++ b/lib/zendesk_apps_support/validations/manifest.rb
@@ -1,4 +1,3 @@
-require 'json'
 require 'uri'
 
 module ZendeskAppsSupport