--- a/zendesk_apps_support.gemspec
+++ b/zendesk_apps_support.gemspec
@@ -13,7 +13,6 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency 'i18n'
   s.add_runtime_dependency 'sass'
   s.add_runtime_dependency 'json'
-  s.add_runtime_dependency 'json-stream'
   s.add_runtime_dependency 'image_size'
   s.add_runtime_dependency 'erubis'
   s.add_runtime_dependency 'jshintrb', '~> 0.3.0'