- Removed `json-stream`: had implemented parallel functionality in NoOverrideHash, since that gem was only used in validating requirements.json let's not force it to every ZAT user.
- Removed require statements from being run before dependencies were needed, especially `sass`

Performance comparison:

### Master

```
ruby -rzendesk_apps_support -e ''  0.33s user 0.12s system 98% cpu 0.454 total
ruby -rzendesk_apps_support -e ''  0.33s user 0.11s system 98% cpu 0.453 total
ruby -rzendesk_apps_support -e ''  0.33s user 0.11s system 98% cpu 0.447 total
ruby -rzendesk_apps_support -e ''  0.35s user 0.12s system 98% cpu 0.482 total
ruby -rzendesk_apps_support -e ''  0.34s user 0.11s system 98% cpu 0.458 total
ruby -rzendesk_apps_support -e ''  0.33s user 0.11s system 98% cpu 0.453 total
ruby -rzendesk_apps_support -e ''  0.37s user 0.14s system 97% cpu 0.521 total
```

### This branch

```
ruby -rzendesk_apps_support -e ''  0.11s user 0.07s system 97% cpu 0.188 total
ruby -rzendesk_apps_support -e ''  0.12s user 0.08s system 97% cpu 0.197 total
ruby -rzendesk_apps_support -e ''  0.11s user 0.07s system 98% cpu 0.190 total
ruby -rzendesk_apps_support -e ''  0.11s user 0.07s system 97% cpu 0.189 total
ruby -rzendesk_apps_support -e ''  0.12s user 0.08s system 96% cpu 0.206 total
```

### Bonus control experiment

```
ruby -e ''  0.09s user 0.06s system 97% cpu 0.154 total
ruby -e ''  0.10s user 0.07s system 97% cpu 0.169 total
ruby -e ''  0.09s user 0.06s system 97% cpu 0.155 total
ruby -e ''  0.10s user 0.06s system 97% cpu 0.161 total
ruby -e ''  0.09s user 0.06s system 97% cpu 0.158 total
```

cc @zendesk/wombat @zendesk/vegemite